online_inference/requirements.txt
requests>=2.25.0
fastapi>=0.47.1
uvicorn>=0.11.7
pandas>=1.1.5
scikit-learn>=0.24.2
marshmallow_dataclass>=8.4.1
pyyaml>=5.4.1
